IGNITE-23114 Code cleanup, javadocs
